Increase LOCK_FACTOR to deal with suspect MongoDB slowness
We are experiencing a MongoDB slowdown due to (it seems) Amazon, when
that happens the recruiter lock expires before the job assignment to
the worker and so it's not able to refresh it and when that happens it
dies. As a fix we are increasing substantially the multiplicator
factor of the maximum time the recruiter can wait between assignments,
since it's an expiration time the worst it can happen is that another
recruiter can take a few seconds more to take leadership.